added Settings Activity
