+++ MockTradeApp/src/main/AndroidManifest.xml
@@ -35,6 +35,15 @@
                 android:value=".MainActivity" />
         </activity>
 
+        <activity
+            android:name=".settings.SettingsActivity"
+            android:label="@string/setting_activity_label" >
+            android:parentActivityName=".MainActivity" >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".MainActivity" />
+        </activity>
+
         <receiver
             android:name=".receivers.BootReceiver"
             android:permission="android.permission.RECEIVE_BOOT_COMPLETED" >
