+++ MockTradeApp/src/main/java/com/balch/mocktrade/settings/SettingsActivity.java
@@ -1,0 +1,96 @@
+/*
+ * Author: Balch
+ * Created: 9/4/14 12:26 AM
+ *
+ * This file is part of MockTrade.
+ *
+ * MockTrade is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MockTrade is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MockTrade.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Copyright (C) 2014
+ */
+
+package com.balch.mocktrade.settings;
+
+import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+import android.preference.PreferenceManager;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+
+import com.balch.android.app.framework.model.ModelFactory;
+import com.balch.mocktrade.R;
+import com.balch.mocktrade.finance.FinanceModel;
+import com.balch.mocktrade.model.ModelProvider;
+
+public class SettingsActivity extends AppCompatActivity
+        implements SharedPreferences.OnSharedPreferenceChangeListener {
+
+
+    public static Intent newIntent(Context context) {
+        return new Intent(context, SettingsActivity.class);
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        setContentView(R.layout.settings_view);
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.settings_view_toolbar);
+        setSupportActionBar(toolbar);
+
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) {
+            actionBar.setDisplayHomeAsUpEnabled(true);
+            actionBar.setDisplayShowHomeEnabled(true);
+        }
+
+        getFragmentManager().beginTransaction().replace(R.id.settings_view_content_frame, new SettingsPreferenceFragment()).commit();
+    }
+
+    @Override
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
+        if (key.equals(getString(R.string.poll_interval))) {
+            ModelFactory modelFactory = ((ModelProvider)this.getApplication()).getModelFactory();
+            FinanceModel financeModel = modelFactory.getModel(FinanceModel.class);
+            financeModel.setQuoteServiceAlarm();
+        }
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        PreferenceManager.getDefaultSharedPreferences(this).registerOnSharedPreferenceChangeListener(this);
+    }
+
+    @Override
+    public void onPause() {
+        PreferenceManager.getDefaultSharedPreferences(this).unregisterOnSharedPreferenceChangeListener(this);
+        super.onPause();
+    }
+
+    public static class SettingsPreferenceFragment extends PreferenceFragment
+    {
+        @Override
+        public void onCreate(final Bundle savedInstanceState)
+        {
+            super.onCreate(savedInstanceState);
+            addPreferencesFromResource(R.xml.settings_pref_screen);
+        }
+    }
+}
