--- MockTradeApp/src/main/java/com/balch/mocktrade/settings/SettingsFragment.java
@@ -1,61 +1,0 @@
-/*
- * Author: Balch
- * Created: 9/4/14 12:26 AM
- *
- * This file is part of MockTrade.
- *
- * MockTrade is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * MockTrade is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with MockTrade.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Copyright (C) 2014
- */
-
-package com.balch.mocktrade.settings;
-
-import android.content.SharedPreferences;
-import android.os.Bundle;
-
-import com.balch.android.app.framework.model.ModelFactory;
-import com.balch.android.app.framework.preference.PreferenceFragment;
-import com.balch.mocktrade.R;
-import com.balch.mocktrade.finance.FinanceModel;
-import com.balch.mocktrade.model.ModelProvider;
-
-public class SettingsFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener {
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.settings_pref_screen);
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-        if (key.equals(getString(R.string.poll_interval))) {
-            ModelFactory modelFactory = ((ModelProvider)this.getActivity().getApplication()).getModelFactory();
-            FinanceModel financeModel = modelFactory.getModel(FinanceModel.class);
-            financeModel.setQuoteServiceAlarm();
-        }
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        getPreferenceManager().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
-    }
-
-    @Override
-    public void onPause() {
-        getPreferenceManager().getSharedPreferences().unregisterOnSharedPreferenceChangeListener(this);
-        super.onPause();
-    }
-}
