+++ MockTradeApp/src/main/res/drawable-xhdpi/ic_refresh.png
